backend/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.6
django-cors-headers==4.8.0
django-extensions==4.1
django-filter==25.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.27.2
psutil==6.0.0
et_xmlfile==2.0.0
kombu==5.5.4
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
qrcode==8.2
redis==6.4.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
